agentcore_memory/pyproject.toml
[project]
name = "agentcore-memory"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aws-opentelemetry-distro>=0.14.0",
    "bedrock-agentcore-starter-toolkit>=0.1.33",
    "boto3>=1.40.74",
    "strands-agents>=1.16.0",
    "strands-agents-tools>=0.2.15",
]
